convert int, unsigned to NSInteger, NSUInteger

+++ DYCreeveyBrowser.h
@@ -23,7 +23,7 @@
 
 
 @interface NSObject (DYCreeveyBrowserDelegate)
-- (void)browser:(NSBrowser *)sender typedString:(NSString *)s inColumn:(int)column;
+- (void)browser:(NSBrowser *)sender typedString:(NSString *)s inColumn:(NSInteger)column;
 - (void)browserWillSendAction:(NSBrowser *)sender;
 - (void)fakeKeyDown:(NSEvent *)e; // for redirecting pgup/down
 @end
