+++ DYVersChecker.h
@@ -15,7 +15,7 @@
 
 @interface DYVersChecker : NSObject {
 	BOOL notify;
-	int responseCode;
+	NSInteger responseCode;
 	NSMutableData *receivedData;
 }
 - initWithNotify:(BOOL)newNotify; // returns nil if can't open connection
